% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{actvector}
\alias{actvector}
\title{Crear Vector de Activación (actvector)}
\usage{
actvector(grid, col.element = 1)
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{col.element}{column number corresponding to the element from which the
activation vector is extracted.}
}
\value{
Returns a vector containing the standardised weights for each of the
constructs associated an element.
}
\description{
Function that extracts a column vector from a RepGrid and
standardises it so that it can be used as an activation vector in a Fuzzy
Cognitive Map.
}
