% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{auc_index}
\alias{auc_index}
\title{PCSD AUC Index -- auc_index()}
\usage{
auc_index(grid, imp, ideal = dim(grid)[2], ...)
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{ideal}{Column number representing the position of the Ideal-Self in the
RepGrid. By default the last column of the RepGrid is set.}

\item{...}{This function inherits all the parameters of \code{\link{pcsd}}
function.}
}
\description{
This function calculates the area under the PCSD curve for each
construct.
}
