% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{CentClose}
\alias{CentClose}
\title{Cálculo de la centralidad a través de la cercanía (CentClose)}
\usage{
CentClose(imp, norm = TRUE)
}
\arguments{
\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importIMP}}.}
}
\value{
Devuelve un vector con el índice de cercanía de cada uno de los
constructos.
}
\description{
Función que permite calcular la cercanía de un constructo
del resto de constructos del mapa cognitivo borroso.
}
\examples{

}
