% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{close_index}
\alias{close_index}
\title{Closeness index -- close_index()}
\usage{
close_index(imp, norm = TRUE)
}
\arguments{
\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{norm}{If TRUE, the values will be standardized. Default is TRUE.}
}
\value{
Returns a vector with the closeness index for each of the
constructs.
}
\description{
Function to calculate the closeness of a construct to the rest
of the constructs within the digraph.
}
