% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docs.R
\name{digraphlayouts}
\alias{digraphlayouts}
\title{Digraph Layouts}
\description{
The GridFCM package has five layouts to represent the digraphs:

\strong{Graphopt:} Closer two constructs are, more influence they have on
each other. To use this layout we must set the layout parameter to "graphopt".

\strong{Multidimensional Scaling:} This layout establishes the coordinates of
the vertices through a multidimensional scaling of the implication matrix.
Closer two constructs are, more similarity there is between them in terms of
their implications. To use this layout we have to set the parameter layout to
"mds".

\strong{Reingold-tilford:} This layout sets the coordinates of the vertices
following the algorithm proposed by Reingold and Tilford. To use this layout,
we can set the layout parameter to "rtcircle" if we want it in circular
format, or to "tree" if we want it in tree format.

\strong{Circle:} This layout arbitrarily places the vertices in a circle
 without any additional meaning. It is useful to represent a clean image and
 avoid overlaps. To use this layout we must set the layout attribute to
 "circle".

\strong{Grid:} This layout arbitrarily places the vertices in a grid
 without any additional meaning. It is useful to represent a clean image and
 avoid overlaps. To use this layout we must set the layout attribute to
 "grid".
}
