% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{DistMatrix}
\alias{DistMatrix}
\title{Cálculo de distancia entre constructos en el digrafo (DistMatrix)}
\usage{
DistMatrix(imp)
}
\arguments{
\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importIMP}}.}
}
\value{
Devuelve la matriz de distancia del digrafo. Matriz que contiene las
distancias de los caminos más cortos de un constructo a otro.
}
\description{
Función que permite calcular la distancia mas corta entre dos
constructos en el digrafo.
}
\examples{

}
