% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docs.R
\docType{data}
\name{elsa2022.imp}
\alias{elsa2022.imp}
\title{Elsa ImpGrid}
\format{
A S4 repgrid object
}
\usage{
data(elsa2022)
}
\description{
A dataset containing Elsa's ImpGrid. From an unpublished case study. You can
use it to test the different functions of the package.
}
\keyword{datasets}
