% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{fcmdigraph}
\alias{fcmdigraph}
\title{Fuzzy Cognitive Map Digraph -- fcmdigraph()}
\usage{
fcmdigraph(
  grid,
  imp,
  results = fcminfer(grid, imp)$values,
  ideal = dim(grid)[2],
  niter = 1,
  layout = "graphopt",
  edge.width = 1.5,
  vertex.size = 1,
  legend = FALSE
)
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{results}{Iteration matrix calculated with
\code{\link{fcminfer}}. By default they are calculated with the Self-Now}

\item{ideal}{Column number representing the position of the Ideal-Self in the
RepGrid. The default is the last column of the RepGrid.}

\item{niter}{Scenario vector to be represented in the digraph expressed by
the row number in the iteration matrix. The default is the first row.}

\item{layout}{Layout to represent the FCM digraph. More information about
layouts in ?\code{\link{digraphlayouts}}.}

\item{edge.width}{Edge width scalar.}

\item{vertex.size}{Vertex size scalar.}

\item{legend}{Draws a legend if TRUE.}
}
\value{
Returns a graphical representation of a digraph of a FCM
}
\description{
This function draws a digraph of the Fuzzy Cognitive Map of an
individual's construct system using a Repgrid and Impgrid.
}
