% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{fcmdigraph3D}
\alias{fcmdigraph3D}
\title{3D Fuzzy Cognitive Digraph -- fcmdigraph3D()}
\usage{
fcmdigraph3D(
  grid,
  imp,
  results = fcminfer(grid, imp)$values,
  ideal = dim(grid)[2],
  niter = 1,
  edge.width = 2,
  vertex.size = 1
)
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{results}{Iterarion matrix calculated with
\code{\link{fcminfer}}. By default they are calculated with the Self-Now}

\item{ideal}{Column number representing the position of the Ideal-Self in the
RepGrid. The default is the last column of the RepGrid.}

\item{niter}{Scenario vector to be represented in the digraph expressed by
the row number in the iteration matrix. The default is the first row.}

\item{edge.width}{Edge width scalar.}

\item{vertex.size}{Vertex size scalar.}
}
\value{
Returns a rgl output of a FCM digraph in a 3D multidimensional
scaling.
}
\description{
Function that draws a three-dimensional FCM digraph of an
personal construct system using a grid technique and an implication
grid. It is represented with a 3D multidimensional scaling.
}
