% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{fcminfer}
\alias{fcminfer}
\title{Inferencia de Escenarios Futuros -- fcminfer()}
\usage{
fcminfer(
  grid,
  imp,
  act.vec = actvector(grid),
  ideal = dim(grid)[2],
  infer = "mk",
  thr = "t",
  lambda = 1,
  iter = 30
)
}
\arguments{
\item{grid}{RepGrid del sujeto desde donde queremos realizar las inferencias.
Debe de ser un objeto importando con la función \code{\link{importgrid}}.}

\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importimp}}.}

\item{act.vec}{Vector de activación creado a través de
\code{\link{actvector}}. Si no se define se utiliza el vector correspondiente
 al primer elemento de la RepGrid.}

\item{ideal}{Posición del ideal dentro de la RepGrid expresado a través del
número de la columna donde se encuentra. Por defecto se establece el último
elemento de la RepGrid.}

\item{infer}{Función de Propagación para la inferencia de escenarios.
Puedes mirar la información sobre las diferentes funciones de propagación
escribiendo ?\code{\link{PropFunctions}}. Por defecto se utiliza la
modificada de Kosko ("mk").}

\item{thr}{Función de Umbral para la inferencia de escenarios. Puedes mirar
la información sonbre las distintas funciones umbral escribiendo
?\code{\link{ThrFunctions}}. Por defecto se utiliza la Tangente Hiperbólica
("t").}

\item{lambda}{Valor de lambda de la función umbral. Solo se aplica si se
escoge función umbral sigmoidal o tangente hiperbólica.}

\item{iter}{Número de iteraciones que se desea inferir}
}
\value{
Devuelve una lista con dos entradas. La entrada $values contiene por
filas cada uno de los vectores de escenario en función del número de
iteraciones. Y la entrada $convergence contiene el número de la iteración
donde se estabilizada el mapa cognitivo borroso.
}
\description{
Función para inferir escenarios futuros simulados del mapa
cognitivo borroso derivados de las relaciones de causalidad.
}
