% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{fcmreport}
\alias{fcmreport}
\title{GridFCM Analisys Report -- fcmreport())}
\usage{
fcmreport(
  x,
  imp,
  name = "report",
  dir = getwd(),
  output = "html",
  edit = FALSE
)
}
\arguments{
\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importimp}}.}

\item{name}{Nombre del archivo de salida}

\item{dir}{Directorio dónde se va a crear el archivo de salida. Por defecto
se estable el directorio de trabajo.}

\item{output}{Work in progress}

\item{edit}{Parametro booleano para indicar si se desea editar el código
fuente del informe. Por defecto se encuentra definido como FALSE.}

\item{grid}{RepGrid del sujeto del que queremos dibujar el Mapa Cognitivo
Borroso. Debe de ser un objeto importado con la función
\code{\link{importgrid}}.}
}
\value{
Devuelve un archivo html que muestra un análisis completo sobre la
RepGrid y la ImpGrid del sujeto.
}
\description{
Función que nos exporta un análisis completo sobre un sujeto
utilizando
}
