% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{FuzzyInfer}
\alias{FuzzyInfer}
\title{Inferencia de Escenarios Futuros (FuzzyInfer)}
\usage{
FuzzyInfer(
  x,
  imp,
  act.vec,
  ideal = dim(x)[2],
  infer = "mk",
  thr = "t",
  lambda = 1,
  iter = 30,
  graph = TRUE
)
}
\arguments{
\item{x}{Rejilla del sujeto desde donde queremos realizar las inferencias.
Debe de ser un objeto de la clase repgrid del paquete OpenRepGrid.}

\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importIMP}}.}

\item{act.vec}{Vector de activación creado a través de
\code{\link{ActVector}}.}

\item{ideal}{Posición del ideal dentro de la rejilla expresado a través del
número de la columna donde se encuentra.}

\item{infer}{Función de Propagación para la inferencia de escenarios.
Información sobre las diferentes funciones de propagación escribiendo
?\code{\link{PropFunctions}}.}

\item{thr}{Función Umbral para la inferencia de escenarios.}

\item{lambda}{Valor de lambda de la función umbral.}

\item{iter}{Número de iteraciones que se desea inferir}

\item{graph}{Devuelve un Gráfico de Comportamiento vs. Tiempo cuando es
\code{TRUE}}
}
\value{
Devuelve una lista (list) que contiene por filas cada uno de los
vectores de esenario en función del número de iteraciones.
}
\description{
Función que infiere los cambios dentro un mapa cognitivo borroso
y lo expresa iteración a iteración. Permite la representación de un Gráfico
de Comportamiento vs. Tiempo (GCT).
}
\examples{

}
