% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{FuzzyMap3D}
\alias{FuzzyMap3D}
\title{Digrafo del Mapa Cognitivo Borroso en 3D (FuzzyMap3D)}
\usage{
FuzzyMap3D(
  x,
  imp,
  results = FuzzyInfer(x, imp, graph = FALSE),
  ideal = dim(x)[2],
  niter = 30,
  edge.width = 2,
  vertex.size = 1
)
}
\arguments{
\item{x}{Rejilla del sujeto del que queremos dibujar el Mapa Cognitivo
Borroso. Debe de ser un objeto de la clase repgrid del paquete OpenRepGrid.}

\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importIMP}}.}

\item{results}{inferencia de escenarios creada con \code{\link{FuzzyInfer}}.}

\item{ideal}{Posición del ideal dentro de la rejilla expresado a través del
número de la columna donde se encuentra.}

\item{niter}{Selección del vector de escenario que se quiere representar en
el mapa. Expresado a través del número de iteraciones.}

\item{edge.width}{Escalar del grosor de las aristas.}

\item{vertex.size}{Escalar del tamaño de los vértices.}
}
\value{
Devuelve una representación gráfica de un digrafo de un Mapa
Cognitivo Borroso en un escalado multidimensional de 3 dimensiones.
}
\description{
Función que nos dibuja un digrafo del Mapa Cognitivo Borroso en
tres dimensiones del sistema de constructos de un individuo a través de una
técnica de rejilla y una rejilla de implicaciones. Utilizando un layout de
escalado multidimensional.
}
\examples{


}
