% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docs.R
\docType{package}
\name{GridFCM-package}
\alias{GridFCM}
\alias{GridFCM-package}
\title{GRID FCM}
\description{
R package for applying Fuzzy Cognitive Maps to Personal
Construct Psychology.

\strong{Logistic Functions:}

\code{\link{importgrid}}

\code{\link{importimp}}

\code{\link{templateimp}}

\strong{Calc Functions:}

\code{\link{actvector}}

\code{\link{dismatrix}}

\code{\link{fcminfer}}

\code{\link{pcsd_summary}}

\strong{Plot Functions:}

\code{\link{fcmdigraph}}

\code{\link{fcmdigraph3D}}

\code{\link{idealdigraph}}

\code{\link{pcsd}}

\code{\link{pcsd_derivative}}

\strong{Index Functions:}

\code{\link{density_index}}

\code{\link{degree_index}}

\code{\link{close_index}}

\code{\link{betw_index}}

\code{\link{auc_index}}

\code{\link{stability_index}}
}
\author{
Alejandro Sanfeliciano -- sanfeliciano.alejandro@gmail.com
}
