% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{idealdigraph}
\alias{idealdigraph}
\title{Ideal Map Digraph -- idealdigraph()}
\usage{
idealdigraph(
  grid,
  imp,
  ideal = dim(grid)[2],
  inc = FALSE,
  layout = "circle",
  edge.width = 1,
  vertex.size = 1,
  legend = FALSE
)
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{ideal}{Column number representing the position of the Ideal-Self in the
RepGrid. The default is the last column of the RepGrid.}

\item{inc}{Shows only inconsistencies (inverse relationships) if TRUE.
Default is FALSE.}

\item{layout}{Layout to represent the FCM digraph. More information about
layouts in ?\code{\link{digraphlayouts}}.}

\item{edge.width}{Edge width scalar.}

\item{vertex.size}{Vertex size scalar.}

\item{legend}{Draws a legend if TRUE. Default is FALSE.}
}
\value{
Returns Ideal FCM digraph.
}
\description{
This function draws an ideal FCM digraph allows to see
implications between Ideal-Self poles.
}
