% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{IdealMap}
\alias{IdealMap}
\title{Digrafo del Yo-Ideal (IdealMap)}
\usage{
IdealMap(
  x,
  imp,
  ideal = dim(x)[2],
  inc = FALSE,
  layout = "circle",
  edge.width = 1,
  vertex.size = 1,
  legend = FALSE
)
}
\arguments{
\item{x}{Rejilla del sujeto del que queremos dibujar el Mapa Cognitivo
Borroso. Debe de ser un objeto de la clase repgrid del paquete OpenRepGrid.}

\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importIMP}}.}

\item{ideal}{Posición del ideal dentro de la rejilla expresado a través del
número de la columna donde se encuentra.}

\item{inc}{Si lo establecemos como TRUE nos muestra solamente las relaciones
de inconsistencia dentro del ideal.}

\item{layout}{Layout que se quiere utilizar para representar el mapa.}

\item{edge.width}{Escalar del grosor de las aristas.}

\item{vertex.size}{Escalar del tamaño de los vértices.}

\item{legend}{Dibuja una leyenda con TRUE.}
}
\value{
Devuelve una representación gráfica de un digrafo del Yo-Ideal del
individuo
}
\description{
Función que nos dibuja un digrafo del Ideal del individuo que
nos permite ver las implicaciones entre los polos del Yo-Ideal.
}
\examples{

}
