% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gridfunctions.R
\name{importgrid}
\alias{importgrid}
\title{Import RepGrid -- importgrid()}
\usage{
importgrid(path, ...)
}
\arguments{
\item{path}{xlsx file path.}

\item{...}{This function inherits all the parameters of function
\code{\link{importExcel}} from OpenRepGrid package.}
}
\value{
Returns an S4 repgrid object containing the data from the RepGrid.
}
\description{
Function used to read RepGrid inside an xlsx file.
}
