% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gridfunctions.R
\name{importimp}
\alias{importimp}
\title{Import ImpGrid -- importimp()}
\usage{
importimp(path, ...)
}
\arguments{
\item{path}{xlsx file path}

\item{...}{This function inherits all the parameters of function
\code{\link{importExcel}} from OpenRepGrid package.}
}
\value{
Returns an S4 repgrid object containing the data from the ImpGrid
}
\description{
Function used to read ImpGrid inside an xlsx file. For the
creation of the ImpGrid template you can use the function
\code{\link{templateimp}}
}
