% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{inc_index}
\alias{inc_index}
\title{Ideal Inconsitencies -- inc_index()}
\usage{
inc_index(grid, imp, ideal = dim(grid)[2])
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{ideal}{Column number representing the position of the Ideal-Self in the
RepGrid. By default the last column of the RepGrid is set.}
}
\value{
WIP.
}
\description{
WIP
}
