% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{pcsd}
\alias{pcsd}
\title{Personal Constructs System Dynamics plot -- pcsd()}
\usage{
pcsd(x, imp, ideal = dim(x)[2], ...)
}
\arguments{
\item{x}{RepGrid sobre la que queremos realizar nuestro PCSD. Debe de haber
sido importada a través de \code{\link{importgrid}}.}

\item{imp}{Impgrid sobre la que queremos realizar nuestro PCSD. Debe de haber
sido importada a través de \code{\link{importimp}}.}

\item{ideal}{Columna donde se encuentra el Yo-Ideal en la Repgrid. Por
defecto se estable la última columna de la Repgrid.}

\item{...}{Esta función hereda todos los parámetos de la función
\code{\link{fcminfer}}}
}
\value{
Gráfico interactivo confeccionado con plotly.
}
\description{
Gráfico que nos permite observar la dinámica del sistema de
constructros personales. Cruza la iteraciones matemáticas de
\code{\link{fcminfer}} con la distancia respecto al ideal de cada uno de los
constructos personales.
}
