% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{pcsd_derivative}
\alias{pcsd_derivative}
\title{PCSD derivative -- pcsd_derivative()}
\usage{
pcsd_derivative(grid, imp, ideal = dim(grid)[2], ...)
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{ideal}{Column number representing the position of the Ideal-Self in the
RepGrid. By default the last column of the RepGrid is set.}

\item{...}{This function inherits all the parameters of \code{\link{pcsd}}
function.}
}
\value{
Return a plot create via plotly r-package.
}
\description{
This function represents the first derivative for each of the
PCSD curves.
}
