% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{pcsd_summary}
\alias{pcsd_summary}
\title{PCSD summary -- pcsd_summary()}
\usage{
pcsd_summary(grid, imp, ideal = dim(grid)[2], ...)
}
\arguments{
\item{grid}{Repgrid del sujeto. Debe de ser un objeto importando con la
función \code{\link{importgrid}}.}

\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importimp}}.}

\item{ideal}{Posición del ideal dentro de la rejilla expresado a través del
número de la columna donde se encuentra. Por defecto se estable el último
elemento de la Repgrid.}

\item{...}{Esta función hereda todos los parámetos de la función
\code{\link{pcsd}}}
}
\value{
Devuelve una matriz con el resumen del PCSD.
}
\description{
Esta función nos devuelve un resumen sobre el PCSD. Nos informa
sobre el valor inicial y final de cada constructo y sobre la diferencia entre
ambos.
}
