% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WImpGrid.R
\name{scenariomatrix}
\alias{scenariomatrix}
\title{Scenario Matrix}
\usage{
scenariomatrix(
  wimp,
  act.vector,
  max.iter = 30,
  e = 1e-04,
  stop.iter = 3,
  force.convergence = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\description{
Scenario Matrix
}
