% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WImpGrid.R
\name{selfdigraph}
\alias{selfdigraph}
\title{Wimp digraph}
\usage{
selfdigraph(
  wimp,
  act.vector,
  scene.matrix = "default",
  niter = 1,
  layout = "graphopt",
  edge.width = 1.5,
  vertex.size = 1,
  legend = FALSE
)
}
\arguments{
\item{legend}{}
}
\description{
Wimp digraph
}
