% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{stability_index}
\alias{stability_index}
\title{PCSD Stability Index -- stability_index()}
\usage{
stability_index(grid, imp, ideal = dim(grid)[2], ...)
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{ideal}{Column number representing the position of the Ideal-Self in the
RepGrid. By default the last column of the RepGrid is set.}

\item{...}{This function inherits all the parameters of \code{\link{pcsd}}
function.}
}
\value{
Returns a vector with the standard deviation of each of the
constructs.
}
\description{
This function returns the standard deviation for each
construct over the mathematical iterations of the PCSD.
}
