% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gridfunctions.R
\name{templateimp}
\alias{templateimp}
\title{Create ImpGrid Template -- templateimp()}
\usage{
templateimp(grid, name = "ImpGrid_Template")
}
\arguments{
\item{grid}{Técnica de Rejilla importada a través de la función
\code{\link{importgrid}}. Debe de ser un objeto RepGrid.}

\item{name}{Nombre para el archivo de salida.}
}
\value{
Exporta un documento xlsx en el directorio de trabajo asignado de la
sesión.
}
\description{
Función que crea en el directorio de trabajo un archivo de excel
con una plantilla personalizada de la Rejilla de Implicaciones teniendo como
base una RepGrid.
}
