% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docs.R
\name{thrfunctions}
\alias{thrfunctions}
\title{Types of Threshold Function}
\description{
The GridFCM package allows for four threshold functions, two of which are
discrete and the other two are continuous:

\strong{Discrete Functions}

\emph{\strong{Bivalent function:}} The bivalent function returns a value of 1
 to the vertex when the incoming activation is greater than 0, and returns 0
 when the incoming activation is less than or equal to 0. This leads to the
 vertices taking discrete and dichotomous values (True or false). To use this
 function we must write as parameter `thr = "b"` in the function
 code{link{fcminfer}}.

\emph{\strong{Trivalent function:}} The trivalent function returns a value of
 1 to the vertex when the incoming activation is greater than 0, returns 0
 when that activation is equal to 0, and returns -1 if that activation is
 less than 0. This leads to the vertices to take discrete values, and unlike
 bivalent, returns three different possible values; useful for discrete
 bipolar dimensions.To use this function, we must write `thr = "tri"` as a
 parameter to the function code{link{fcminfer}}.

\strong{Continuous Functions}

\emph{\strong{Sigmoidal Function:}} The sigmoidal function is the continuous
version of the bivalent function. Forces the vertex to be set to an
activation range between 0 and 1. If we use this function we must specify an
associated lambda which defines the discriminating power of the function;
high values of lambda bring the sigmoidal function closer to a bivalent
function, low values bring it closer to a linear function. To use this
function we must write as parameter `thr = "s"` in the function
\code{link{fcminfer}}.

\emph{\strong{Hyperbolic Tangent Function:}} The hyperbolic tangent function
is the continuos version of the trivalent function. Forces the vertez to be
set to an activation range between -1 and +1. If we use this function
we must specify an associated lambda which defines the discriminating power
of the function; high values of lambda bring the Hyperbolic Tangent Function
closer to a trivalent function, low values bring it closer to a linear
function. To use this function we must write as parameter `thr = "t"` in the
function \code{link{fcminfer}}.
}
