% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{actvector}
\alias{actvector}
\title{Crear Vector de Activación (actvector)}
\usage{
actvector(x, col.element = 1)
}
\arguments{
\item{x}{Rejilla del sujeto desde donde queremos extraer el vector de
activación. Debe de ser un objeto de la clase repgrid del paquete
OpenRepGrid.}

\item{col.element}{Elemento desde el que extraemos el vector de activación.
Por defecto se establece el primer elemento, que coincide normalmente con el
Yo-Actual.}
}
\value{
Devuelve un vector que contiene los pesos para cada uno de los
constructos asociados a un elemento.
}
\description{
Función que extrae un vector para un elemento de la rejilla y lo
transforma en un vector de activación apropiado para utilizarse a la hora de
crear un mapa cognitivo borroso.
}
\examples{

}
