% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{betw_index}
\alias{betw_index}
\title{Betweeness index -- betw_index()}
\usage{
betw_index(imp, norm = TRUE)
}
\arguments{
\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{norm}{If TRUE, the values will be standardized. Default is TRUE.}
}
\value{
Returns a vector with the betweeness index for each of the
constructs.
}
\description{
Function that calculates the betweenness of each of the
constructs. This is the number of times a geodesic path (shortest path)
between two other constructs passes through that construct in the digraph.
}
