% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{betw_index}
\alias{betw_index}
\title{Cálculo de la centralidad a través de la intermediación (CentBetw)}
\usage{
betw_index(imp, norm = TRUE)
}
\arguments{
\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importIMP}}.}

\item{norm}{Si es TRUE devuelve la intermediación de los constructos
normalizada en función del número de vértices. Por defecto se establece en
TRUE.}
}
\value{
Devuelve un vector con el índice de intermediación de cada uno de los
constructos.
}
\description{
Función que permite calcular la intermediación de un
constructo. Esto es el número de veces que un camino geodésico entre otros
dos constructos pasa por dicho constructo.
}
\examples{

}
