% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{close_index}
\alias{close_index}
\title{Closeness index -- close_index()}
\usage{
close_index(imp, norm = TRUE)
}
\arguments{
\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importimp}}.}

\item{norm}{Si es TRUE devuelve la cercanía de los constructos normalizada en
función del número de vértices. Por defecto se establece en TRUE.}
}
\value{
Devuelve un vector con el índice de cercanía de cada uno de los
constructos.
}
\description{
Función que permite calcular la cercanía de un constructo
del resto de constructos del mapa cognitivo borroso.
}
