% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{degree_index}
\alias{degree_index}
\title{Degree Index -- degree_index()}
\usage{
degree_index(imp, method = "simple")
}
\arguments{
\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importimp}}.}

\item{method}{Método para calcular el grado de centralidad.}
}
\value{
Devuelve una lista con los datos de centralidad por constructo y
separados por grado de entrada y grado de salida.
}
\description{
Función que permite calcular la centralidad de los constructos
teniendo en cuenta su grado. Entendiendo este como el grado de conexión que
mantiene con el resto de constructos.
}
