% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{density_index}
\alias{density_index}
\title{Cálculo de la Densidad del Mapa (IMPdensity)}
\usage{
density_index(imp)
}
\arguments{
\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importIMP}}.}
}
\value{
Devuelve una valor de 0 a 1 que representa la proporcion de aristas
existentes en el grafo sobre el número máximo de aristas posibles.
}
\description{
Función que permite calcular la densidad de aristas dentro del
grafo que genera la matriz de implicaciones.
}
\examples{

}
