% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docs.R
\name{digraphlayouts}
\alias{digraphlayouts}
\title{Tipos de layouts para el digrafo}
\description{
El paquete GridFCM permite cinco layouts para el digrafo:

\strong{Graphopt:} Este layout representa el digrafo creando clusters en función de la influencia que tienen los constructos entre si. Cuanto más cerca estén dos constructos,
mayor influencia hay entre ellos. Para utilizar este layout debemos igualar el atributo layout a "graphopt".

\strong{Multidimensional Scaling:} Este layout establece las coordenadas de los vértices a través de una escalado multidimensional de la matriz de implicaciones.
Cuanto más cerca estén dos constructos, una mayor similitud hay entre ellos en referencia a sus implicaciones. Para utilizar este layout debemos igualar el atributo layout a "mds".

\strong{Reingold-tilford:} Este layout establece las coordenadas de los vértices siguiendo el algoritmo propuesto por Reingold y Tilford. Para utilizar este layout, podemos igualar
el atributo layout a "rtcircle" si lo queremos en formato circular, o a "tree" si lo queremos en formato de arbol.

\strong{Circle:} Este layout coloca los vértices en círculo de forma arbitraria sin ningún significado adicional. Es util para representar una imagen limpia y evitar superposiciones.
Para utilizar este layout debemos igualar el atributo layout a "circle".

\strong{Grid:} Este layout coloca los vértices siguiendo una rejilla de forma arbitraria sin ningún significado adicional. Es util para representar una imagen limpia y evitar
superposiciones. Para utilizar este layout debemos igualar el atributo layout a "grid".
}
