% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{dismatrix}
\alias{dismatrix}
\title{Cálculo de distancia entre constructos en el digrafo (IMPdistances)}
\usage{
dismatrix(imp, mode = "out")
}
\arguments{
\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importIMP}}.}

\item{mode}{Modo de calcular las distancias en función de la dirección de las
aristas. "out" las calculamos respetando la dirección de las aristas, "in"
a través de la inversa de la dirección de las aristas y "all" sin tener en
cuenta la dirección del grafo.}
}
\value{
Devuelve la matriz de distancia del digrafo. Matriz que contiene las
distancias de los caminos más cortos de un constructo a otro.
}
\description{
Función que permite calcular la distancia mas corta entre dos
constructos en el digrafo.
}
\examples{

}
