% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{fcmdigraph}
\alias{fcmdigraph}
\title{Fuzzy Cognitive Map Digraph -- fcmdigraph()}
\usage{
fcmdigraph(
  grid,
  imp,
  results = fcminfer(grid, imp)$values,
  ideal = dim(grid)[2],
  niter = 1,
  layout = "graphopt",
  edge.width = 1.5,
  vertex.size = 1,
  legend = FALSE
)
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{results}{scenario inference values calculated with
\code{\link{fcminfer}}. By default they are calculated with the Self-Now}

\item{ideal}{Column number representing the position of the Ideal-Self in the
RepGrid. The default is the last column of the RepGrid.}

\item{niter}{Selección del vector de escenario que se quiere representar en
el mapa. Expresado a través del número de iteraciones. Por defecto se
establece la primera iteracion.}

\item{layout}{Layout que se quiere utilizar para representar el mapa. Más
información de los layouts con ?\code{\link{digraphlayouts}}.}

\item{edge.width}{Escalar del grosor de las aristas.}

\item{vertex.size}{Escalar del tamaño de los vértices.}

\item{legend}{Dibuja una leyenda con TRUE.}
}
\value{
Devuelve una representación gráfica de un digrafo de un Mapa
Cognitivo Borroso.
}
\description{
This function draws a digraph of the Fuzzy Cognitive Map of an
individual's construct system using a Repgrid and Impgrid.
}
