% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{fcminfer}
\alias{fcminfer}
\title{Inferencia de Escenarios Futuros -- fcminfer()}
\usage{
fcminfer(
  grid,
  imp,
  act.vec = actvector(grid),
  ideal = dim(grid)[2],
  infer = "mk",
  thr = "t",
  lambda = 1,
  iter = 30
)
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{act.vec}{Activation vector created via \code{\link{actvector}} function.
The default vector is the first element of the RepGrid.}

\item{ideal}{Column number representing the position of the Ideal-Self in the
RepGrid. By default the last column of the RepGrid is set.}

\item{infer}{Propagation function for scenario inference. More information in
?\code{\link{propfunctions}}.}

\item{thr}{Threshold function for scenario inference. More information in
?\code{\link{thrfunctions}}.}

\item{lambda}{Lambda value of the threshold function. Only applicable in
sigmoidal or hyperbolic tangent threshold function}

\item{iter}{Number of iterations to infer.}
}
\value{
Return a list with two entries. The $values entry contains in rows
each of the scenario vectors according to the number of iterations. And the
$convergence entry contains the number of the iteration where the
Fuzzy Cognitive Map is stabilised.
}
\description{
Function to infer simulated future scenarios from a RepGrid and
an Impgrid.
}
