% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{fcminfer}
\alias{fcminfer}
\title{Inferencia de Escenarios Futuros -- fcminfer()}
\usage{
fcminfer(
  x,
  imp,
  act.vec = actvector(x),
  ideal = dim(x)[2],
  infer = "mk",
  thr = "t",
  lambda = 1,
  iter = 30
)
}
\arguments{
\item{x}{Rejilla del sujeto desde donde queremos realizar las inferencias.
Debe de ser un objeto de la clase repgrid del paquete OpenRepGrid.}

\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importIMP}}.}

\item{act.vec}{Vector de activación creado a través de
\code{\link{ActVector}}.}

\item{ideal}{Posición del ideal dentro de la rejilla expresado a través del
número de la columna donde se encuentra.}

\item{infer}{Función de Propagación para la inferencia de escenarios.
Puedes mirar la información sobre las diferentes funciones de propagación
escribiendo ?\code{\link{PropFunctions}}. Por defecto se utiliza la
modificada de Kosko ("mk").}

\item{thr}{Función Umbral para la inferencia de escenarios. Puedes mirar la
información sonbre las distintas funciones umbral escribiendo
?\code{\link{ThrFunctions}}. Por defecto se utiliza la Tangente Hiperbólica
("t").}

\item{lambda}{Valor de lambda de la función umbral. Solo se aplica si se
escoge función umbral sigmoidal o tangente hiperbólica.}

\item{iter}{Número de iteraciones que se desea inferir}
}
\value{
Devuelve una lista con dos entradas. La entrada $values contiene por
filas cada uno de los vectores de escenario en función del número de
iteraciones. Y la entrada $convergence contiene el número de la iteración
donde se estabilizada el mapa cognitivo borroso.
}
\description{
Función que infiere los cambios dentro un mapa cognitivo borroso
y lo expresa iteración a iteración.
}
\examples{

}
