% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{fcmreport}
\alias{fcmreport}
\title{GridFCM Analisys Report -- fcmreport())}
\usage{
fcmreport(
  grid,
  imp,
  name = "report",
  dir = getwd(),
  output = "html",
  edit = FALSE
)
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{name}{Name of the output file.}

\item{dir}{Directory where the output file is to be created. Default is the
working directory.}

\item{output}{Work in progress.}

\item{edit}{Allows editing of the draft if TRUE. Default is FALSE.}
}
\value{
Returns a report with a full GridFCM analysis
}
\description{
A function that exports a complete analysis of a subject using
rmarkdown draft
}
