% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docs.R
\docType{package}
\name{GridFCM-package}
\alias{GridFCM}
\alias{GridFCM-package}
\title{GRID FCM}
\description{
Paquete de R para la creación de mapas cognitivos borrosos a través de la técnica de rejilla
}
\section{Functions}{


\strong{Funciones de Utilidades:}

\code{\link{importGRID}}

\code{\link{importIMP}}

\code{\link{templateIMP}}

\code{\link{ActVector}}

\strong{Funciones de Representación:}

\code{\link{FuzzyInfer}}

\code{\link{FuzzyMap}}

\code{\link{FuzzyMap3D}}

\code{\link{IdealMap}}

\strong{Funciones de Índices:}

\code{\link{IMPdensity}}

\code{\link{IMPdistances}}

\code{\link{CentDegree}}

\code{\link{CentClose}}

\code{\link{CentBetw}}
}

\author{
Alejandro Sanfeliciano (GICUNED)
}
