% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docs.R
\docType{package}
\name{GridFCM-package}
\alias{GridFCM}
\alias{GridFCM-package}
\title{GRID FCM}
\description{
Paquete de R para la creación de mapas cognitivos borrosos a través de la técnica de rejilla
}
\section{Functions}{

\code{\link{importIMP}}

\code{\link{ActVector}}

\code{\link{FuzzyInfer}}

\code{\link{FuzzyMap}}

\code{\link{FuzzyMap3D}}

\code{\link{IdealMap}}
}

\author{
Alejandro Sanfeliciano (GICUNED)
}
