% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMfunctions.R
\name{idealdigraph}
\alias{idealdigraph}
\title{Ideal Map Digraph -- idealdigraph()}
\usage{
idealdigraph(
  grid,
  imp,
  ideal = dim(grid)[2],
  inc = FALSE,
  layout = "circle",
  edge.width = 1,
  vertex.size = 1,
  legend = FALSE
)
}
\arguments{
\item{grid}{RepGrid del sujeto del que queremos dibujar el Mapa Cognitivo
Borroso. Debe de ser un objeto importado con la función
\code{\link{importgrid}}.}

\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importimp}}.}

\item{ideal}{Posición del ideal dentro de la rejilla expresado a través del
número de la columna donde se encuentra. Por defecto se establece el último
elemento de la rejilla.}

\item{inc}{Si lo establecemos como TRUE nos muestra solamente las relaciones
de inconsistencia (relaciones inversas) dentro del ideal.}

\item{layout}{Layout que se quiere utilizar para representar el mapa.}

\item{edge.width}{Escalar del grosor de las aristas.}

\item{vertex.size}{Escalar del tamaño de los vértices.}

\item{legend}{Dibuja una leyenda con TRUE.}
}
\value{
Devuelve una representación gráfica de un digrafo del Yo-Ideal del
individuo
}
\description{
Función que nos dibuja un digrafo del Ideal del individuo que
nos permite ver las implicaciones entre los polos del Yo-Ideal.
}
