% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gridfunctions.R
\name{importgrid}
\alias{importgrid}
\title{Import RepGrid -- importgrid()}
\usage{
importgrid(path, ...)
}
\arguments{
\item{path}{Ruta donde se encuentra el archivo de excel que contiene la
la técnica de rejilla.}
}
\value{
Devuelve un objeto S4 repgrid del paquete OpenRepGrid que contiene la
información de la técnica de rejilla.
}
\description{
Función que permite leer una técnica de rejilla contenida
dentro de un archivo de excel.
}
