% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gridfunctions.R
\name{importIMP}
\alias{importIMP}
\title{Importar Rejilla de Implicaciones (importIMP)}
\usage{
importIMP(path)
}
\arguments{
\item{path}{Ruta donde se encuentra el archivo de excel que contiene la
rejilla de implicaciones.}
}
\value{
Devuelve una matriz que contiene los datos de la matriz de
implicaciones
}
\description{
Función que permite leer una rejilla de implicaciones contenida
dentro de un archivo de excel.
}
\examples{

}
