% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gridfunctions.R
\name{importimp}
\alias{importimp}
\title{Import ImpGrid -- importimp()}
\usage{
importimp(path, ...)
}
\arguments{
\item{path}{Ruta donde se encuentra el archivo de excel que contiene la
ImpGrid.}

\item{...}{Esta función hereda todos los parámetros de la función
\code{\link{importExcel}} del paquete OpenRepGrid.}
}
\value{
Devuelve un objeto S4 repgrid que contiene los datos de la matriz de
implicaciones
}
\description{
Función que permite leer una ImpGrid contenida
dentro de un archivo xlsx. Para la creación de la plantilla xlsx de la
ImpGrid se puede utilizar la función \code{\link{templateimp}}
}
