% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gridfunctions.R
\name{importimp}
\alias{importimp}
\title{Importar Rejilla de Implicaciones (importimp)}
\usage{
importimp(path, ...)
}
\arguments{
\item{path}{Ruta donde se encuentra el archivo de excel que contiene la
rejilla de implicaciones.}

\item{method}{Metodo que se ha seguido a la hora de aplicar la rejilla de
implicaciones. Debe tomar el valor "hinkle" si se utiliza el método de Hinkle
y "fransella" en el caso del método de fransella.}
}
\value{
Devuelve una matriz que contiene los datos de la matriz de
implicaciones
}
\description{
Función que permite leer una rejilla de implicaciones contenida
dentro de un archivo de excel.
}
\examples{

}
