% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{pcsd_derivative}
\alias{pcsd_derivative}
\title{PCSD derivative -- pcsd_derivative()}
\usage{
pcsd_derivative(x, imp, ideal = dim(x)[2], ...)
}
\arguments{
\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importimp}}.}

\item{ideal}{Posición del ideal dentro de la rejilla expresado a través del
número de la columna donde se encuentra. Por defecto se estable el último
elemento de la Repgrid.}

\item{...}{Esta función hereda todos los parámetos de la función
\code{\link{pcsd}}}

\item{grid}{Repgrid del sujeto. Debe de ser un objeto importando con la
función \code{\link{importgrid}}.}
}
\value{
Devuelve una representación gráfica creada con el paquete plotly.
}
\description{
Esta función reprensenta la primera derivada para cada una de
las curvas del PCSD.
}
