% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{pcsd_summary}
\alias{pcsd_summary}
\title{PCSD summary -- pcsd_summary()}
\usage{
pcsd_summary(grid, imp, ideal = dim(grid)[2], ...)
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{imp}{Subject's ImpGrid. It must be an S4 object imported by the
\code{\link{importimp}} function.}

\item{ideal}{Column number representing the position of the Ideal-Self in the
RepGrid. By default the last column of the RepGrid is set.}

\item{...}{This function inherits all the parameters of \code{\link{pcsd}}
function.}
}
\value{
Returns a matrix with the PCSD summary.
}
\description{
This function returns a summary of the PCSD. It informs us the
initial and final value of each construct and the difference between them.
}
