% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docs.R
\name{propfunctions}
\alias{propfunctions}
\title{Types of Propagation Punctions}
\description{
The GridFCM package allows three propagation functions:

\strong{Kosko's inference:} It consists of a matrix product between the
activation vector and the weight matrix. The new activation of the vertices
depends on the activation of their inputs and the weight of their relations.
o use it in the function \code{link{fcminfer}} the argument infer must be
equal to "k".

\strong{Modified Kosko's inference:} It consists of adding the matrix product
between the activation vector and the weight matrix to the current state of
the vertices. The new activation of the vertices depends on the activation of
their inputs, the weight of their relations and the previous state of the
network. To use it in the function \code{link{fcminfer}} the parameter infer
must be equal to "mk".


\strong{Rescalated inference:} Similar to the Modified Kosko but with a
rescaling of the values. To use it in the function \code{link{fcminfer}}
the parameter infer must be equal to "r".
}
