% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFunctions.R
\name{stability_index}
\alias{stability_index}
\title{PCSD Stability Index -- stability_index()}
\usage{
stability_index(grid, imp, ideal = dim(grid)[2], ...)
}
\arguments{
\item{grid}{Repgrid del sujeto. Debe de ser un objeto importando con la
función \code{\link{importgrid}}.}

\item{imp}{Matriz de implicaciones del sujeto importada con
\code{\link{importimp}}.}

\item{ideal}{Posición del ideal dentro de la rejilla expresado a través del
número de la columna donde se encuentra. Por defecto se estable el último
elemento de la Repgrid.}

\item{...}{Esta función hereda todos los parámetos de la función
\code{\link{pcsd}}}
}
\value{
Devuelve un vector con los valores de la desviación típica para cada
uno de los constructos.
}
\description{
Esta función nos devuelve la desviación típica para cada uno de
los constructos personales a lo largo de la iteraciones matemáticas del PCSD.
}
