% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gridfunctions.R
\name{templateIMP}
\alias{templateIMP}
\title{Importar Técnica de Rejilla (importGRID)}
\usage{
templateIMP(x)
}
\value{
Exporta un documento xlsx en el directorio de trabajo asignado de la
sesión.
}
\description{
Función que crea en el directorio de trabajo un archivo de excel
con una plantilla personalizada de la Rejilla de Implicaciones en base a una
Tecnica de Rejilla
}
\examples{

}
