% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gridfunctions.R
\name{templateimp}
\alias{templateimp}
\title{Create ImpGrid Template -- templateimp()}
\usage{
templateimp(grid, name = "ImpGrid_Template")
}
\arguments{
\item{grid}{Subject's RepGrid. It must be an S4 object imported by the
\code{\link{importgrid}} function.}

\item{name}{Output file name. Default is "ImpGrid_Template"}
}
\value{
Export an xlsx file in the working directory.
}
\description{
Function that creates an xlsx file in the working directory.
This file contains an impgrid template according to the specified repgrid.
}
